refactor: use `usize` for `datalen` parameter in `_zbar_image_scanner_alloc_sym`
The parameter is used for `Vec::reserve_exact` which expects `usize`, eliminating unnecessary casts.

+++ src/img_scanner.rs
@@ -438,7 +438,7 @@ pub unsafe fn _zbar_image_scanner_quiet_border(iscn: &mut zbar_image_scanner_t) 
 pub(crate) unsafe fn _zbar_image_scanner_alloc_sym(
     _iscn: &mut zbar_image_scanner_t,
     sym_type: c_int,
-    datalen: c_int,
+    datalen: usize,
 ) -> *mut zbar_symbol_t {
     let sym = symbol_alloc_zeroed();
 
@@ -451,7 +451,7 @@ pub(crate) unsafe fn _zbar_image_scanner_alloc_sym(
 
     // Reserve capacity for the data (no null terminator needed in Rust)
     if datalen > 0 {
-        sym_ref.data.reserve_exact(datalen as usize);
+        sym_ref.data.reserve_exact(datalen);
     }
 
     sym
@@ -654,7 +654,7 @@ pub unsafe fn symbol_handler(dcode: *mut zbar_decoder_t) {
     }
 
     // Allocate new symbol (no null terminator needed in Rust)
-    let sym = _zbar_image_scanner_alloc_sym(&mut *iscn, type_ as c_int, data.len() as c_int);
+    let sym = _zbar_image_scanner_alloc_sym(&mut *iscn, type_ as c_int, data.len());
     let sym_ref = &mut *sym;
     sym_ref.configs = zbar_decoder_get_configs(&*dcode, type_);
     sym_ref.modifiers = zbar_decoder_get_modifiers(dcode);
@@ -993,7 +993,7 @@ pub unsafe fn _zbar_scan_image(
             // Create composite symbol (no null terminator needed in Rust)
             let datalen = (*ean).data.len() + (*addon).data.len();
             let ean_sym =
-                _zbar_image_scanner_alloc_sym(&mut *iscn, ZBAR_COMPOSITE, datalen as c_int);
+                _zbar_image_scanner_alloc_sym(&mut *iscn, ZBAR_COMPOSITE, datalen);
             (*ean_sym).orient = (*ean).orient;
             (*ean_sym).syms = symbol_set_create();
